backend/requirements.txt
fastapi==0.98.0
uvicorn[standard]
python-multipart
asgi-logger
python-dotenv
motor
bson